Switched core creation logic from deleting core.properties to UNLOAD.
olr won't allow us to CREATE a new core when one with the same
properties already exists. This is problematic when treating
create as an idempotent operation, and moreso when creation fails
for one reason or another, or AAE believes the index needs to be
re-created (i.e. bad schema, missing core.properties, etc...).
Without first unloading the core Solr may believe the core still
exists (i.e. exists in memory, but there's no core.properties on
disk), but AAE may disagree and enter a cycle of continuously
trying to CREATE the core, only to get an already_exists error.
This cycle makes it impossible for AAE to recover Solr from
these states.

To address this, we first unload the core from Solr. This may be
a moderately expensive process, but given we should only be
issuing creation requests when either AAE believes Solr is
missing the index, or we are creating it for the first time, the
impact should be minimal (and somewhat unavoidable; if AAE is
replacing the whole index, that's what it's going to do anyway).